+++ src/main/java/team/cats/psychological/service/SchoolService.java
@@ -5,10 +5,8 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.github.yitter.idgen.YitIdHelper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import team.cats.psychological.entity.Area;
-import team.cats.psychological.entity.School;
-import team.cats.psychological.entity.TeacherSchool;
-import team.cats.psychological.entity.Users;
+import team.cats.psychological.base.BaseException;
+import team.cats.psychological.entity.*;
 import team.cats.psychological.mapper.*;
 import team.cats.psychological.vo.ListItemVo;
 import team.cats.psychological.vo.SchoolView;
@@ -41,18 +39,32 @@ public class SchoolService {
     }
 
     public void delSchool(Long schoolId) {
+        QueryWrapper<Classes> classesQueryWrapper = new QueryWrapper<>();
+        classesQueryWrapper.eq("school_id",schoolId);
+        List<Classes> classes = classesMapper.selectList(classesQueryWrapper);
+        if(classes.size()>0){
+            throw new BaseException(400,"学校已绑定班级无法删除");
+        }
+
+        QueryWrapper<TeacherSchool>  teacherSchoolQueryWrapper = new QueryWrapper<>();
+        teacherSchoolQueryWrapper.eq("school_id",schoolId);
+        List<TeacherSchool> teacherSchools = teacherSchoolMapper.selectList(teacherSchoolQueryWrapper);
+        if (teacherSchools.size()>0){
+            throw new BaseException(400,"学校内还有老师无法删除");
+        }
+
         schoolMapper.deleteById(schoolId);
     }
 
     public List<School> selectSchool() {
-        List<School> schoolList=new ArrayList<>();
+        List<School> schoolList = new ArrayList<>();
 
         long userId = StpUtil.getLoginIdAsLong();
         Users users = usersMapper.selectById(userId);
         Long userRole = users.getUserRole();
         if (userRole == 0) {
             List<School> schools = schoolMapper.selectList(null);
-            schoolList=schools;
+            schoolList = schools;
         } else if (userRole == 1) {
             QueryWrapper<Area> areaQueryWrapper = new QueryWrapper<>();
             areaQueryWrapper.eq("area_principal", userId);
@@ -65,9 +77,9 @@ public class SchoolService {
                     schoolList.add(school);
                 }
             }
-        } else if (userRole==4){
+        } else if (userRole == 4) {
             QueryWrapper<TeacherSchool> teacherSchoolQueryWrapper = new QueryWrapper<>();
-            teacherSchoolQueryWrapper.eq("teacher_id",userId);
+            teacherSchoolQueryWrapper.eq("teacher_id", userId);
             List<TeacherSchool> teacherSchools = teacherSchoolMapper.selectList(teacherSchoolQueryWrapper);
             for (TeacherSchool teacherSchool : teacherSchools) {
                 schoolList.add(schoolMapper.selectById(teacherSchool.getSchoolId()));
@@ -84,7 +96,7 @@ public class SchoolService {
         Users users = usersMapper.selectById(userId);
         Long userRole = users.getUserRole();
         if (userRole == 0) {
-            schoolViewList=schoolList;
+            schoolViewList = schoolList;
         } else if (userRole == 1) {
             QueryWrapper<Area> areaQueryWrapper = new QueryWrapper<>();
             areaQueryWrapper.eq("area_principal", userId);
@@ -95,27 +107,44 @@ public class SchoolService {
                 List<School> schools = schoolMapper.selectList(schoolQueryWrapper);
                 for (SchoolView schoolView : schoolList) {
                     for (School school : schools) {
-                        if (schoolView.getSchoolId().equals(school.getSchoolId())){
+                        if (schoolView.getSchoolId().equals(school.getSchoolId())) {
                             schoolViewList.add(schoolView);
                         }
                     }
                 }
 
             }
-        }else if (userRole==4){
+        } else if (userRole == 4) {
             QueryWrapper<TeacherSchool> teacherSchoolQueryWrapper = new QueryWrapper<>();
-            teacherSchoolQueryWrapper.eq("teacher_id",userId);
+            teacherSchoolQueryWrapper.eq("teacher_id", userId);
             List<TeacherSchool> teacherSchools = teacherSchoolMapper.selectList(teacherSchoolQueryWrapper);
             for (TeacherSchool teacherSchool : teacherSchools) {
                 for (SchoolView schoolView : schoolList) {
-                    if (teacherSchool.getSchoolId().equals(schoolView.getSchoolId())){
+                    if (teacherSchool.getSchoolId().equals(schoolView.getSchoolId())) {
                         schoolViewList.add(schoolView);
                     }
                 }
             }
-        }else if (userRole==2){
-            schoolViewList=schoolList;
+        } else if (userRole == 2) {
+            schoolViewList = schoolList;
         }
+        return schoolViewList.stream().map(
+                i -> {
+                    final List<ListItemVo> children = classesMapper.selectBySchoolId(i.getSchoolId()).stream().map(t -> new ListItemVo()
+                            .setLabel(t.getClassName())
+                            .setValue(t.getClassId())).collect(Collectors.toList());
+                    return new ListItemVo()
+                            .setLabel(i.getSchoolName())
+                            .setValue(i.getSchoolId())
+                            .setChildren(children);
+                }).collect(Collectors.toList());
+    }
+
+    @Transactional
+    public List<ListItemVo> getSchoolListNoLogin() {
+        List<SchoolView> schoolViewList = new ArrayList<>();
+        List<SchoolView> schoolList = schoolMapper.getSchoolList();
+        schoolViewList = schoolList;
         return schoolViewList.stream().map(
                 i -> {
                     final List<ListItemVo> children = classesMapper.selectBySchoolId(i.getSchoolId()).stream().map(t -> new ListItemVo()
